Refactor temp_uploaded_file fuera de test helpers
