+++ infra/apps/catalog/helpers/temp_uploaded_file.py
@@ -1,4 +1,5 @@
 import os
+
 from django.core.files.uploadedfile import TemporaryUploadedFile
