+++ infra/apps/catalog/models/node.py
@@ -3,7 +3,7 @@ from django.db import models
 
 from infra.apps.catalog.exceptions.catalog_not_uploaded_error import CatalogNotUploadedError
 from infra.apps.catalog.storage.paths import latest_json_catalog_path
-from infra.apps.catalog.tests.helpers.temp_uploaded_file import temp_uploaded_file
+from infra.apps.catalog.helpers.temp_uploaded_file import temp_uploaded_file
 from infra.apps.catalog.validator.catalog_data_validator import CatalogDataValidator
